pyproject.toml
[project]
name = "bella-companion"
version = "0.1.7"
description = "Bayesian Evolutionary Layered Learning Architectures (BELLA) companion"
readme = "README.md"
authors = [
    { name = "gabriele-marino", email = "gabmarino.8601@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "arviz>=0.22.0",
    "autoregistry>=1.2.0",
    "dendropy>=5.0.8",
    "dotenv>=0.9.9",
    "kaleido>=1.2.0",
    "phylogenie>=3.1.4",
    "plotly>=6.5.0",
    "plotly-stubs>=0.0.6",
    "scipy>=1.15.3",
    "seaborn>=0.13.2",
    "shap>=0.49.1",
]

[dependency-groups]
dev = [
    "joblib-stubs>=1.5.2.0.20250831",
    "pandas-stubs>=2.3.2.250926",
    "pyright>=1.1.407",
    "scipy-stubs>=1.15.3.0",
]

[tool.pyright]
typeCheckingMode = "strict"

[project.scripts]
bella = "bella.cli:main"

[build-system]
requires = ["uv_build>=0.9.5,<0.10.0"]
build-backend = "uv_build"
